@@ -4,7 +4,7 @@
     <parent>
         <groupId>com.ettrema</groupId>
         <artifactId>clyde</artifactId>
-        <version>2.0.1-SNAPSHOT</version>
+        <version>2.0.2-SNAPSHOT</version>
     </parent>
     <groupId>com.ettrema</groupId>
     <artifactId>clyde-api</artifactId>
@@ -17,8 +17,8 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>2.3.2</version>
                 <configuration>
-                    <source>1.6</source>
-                    <target>1.6</target>
+                    <source>1.7</source>
+                    <target>1.7</target>
                 </configuration>
             </plugin>
         </plugins>
@@ -34,6 +34,11 @@
             <artifactId>milton-json</artifactId>
             <version>${milton.version}</version>
         </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>milton-caldav</artifactId>
+            <version>${milton.version}</version>
+        </dependency>        
         <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>milton-ftp</artifactId>
@@ -114,11 +119,6 @@
             <artifactId>junit</artifactId>
             <version>4.8.1</version>
             <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>milton-caldav</artifactId>
-            <version>1.6.9-SNAPSHOT</version>
         </dependency>
     </dependencies>
 </project>
