@@ -3,7 +3,7 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.ettrema</groupId>
     <artifactId>clyde</artifactId>
-    <version>2.0.1-SNAPSHOT</version>
+    <version>2.0.2-SNAPSHOT</version>
     <packaging>pom</packaging>
     <name>Clyde Parent</name>
     <url>http://clyde.ettrema.com</url>
@@ -11,7 +11,7 @@
         <connection>scm:git:https://github.com/ettrema/clyde.git</connection>
     </scm>
     <properties>
-        <milton.version>1.6.9-SNAPSHOT</milton.version>
+        <milton.version>1.7.0</milton.version>
         <berry.version>1.3</berry.version>
         <geroa.version>1.4</geroa.version>
         <morton.version>1.2</morton.version>
@@ -52,13 +52,15 @@
         <plugins>
             <plugin>
                 <artifactId>maven-compiler-plugin</artifactId>
+                <version>2.3.2</version>
                 <configuration>
-                    <source>1.6</source>
-                    <target>1.6</target>
+                    <source>1.7</source>
+                    <target>1.7</target>
                 </configuration>
             </plugin>
             <plugin>
                 <artifactId>maven-source-plugin</artifactId>
+                <version>2.1.2</version>
                 <executions>
                     <execution>
                         <id>attach-sources</id>
