@@ -1,17 +1,12 @@
-
 package com.ettrema.web;
 
 import com.bradmcevoy.common.Path;
-import com.bradmcevoy.http.Auth;
-import com.bradmcevoy.http.FileItem;
-import com.bradmcevoy.http.PostableResource;
-import com.bradmcevoy.http.Range;
-import com.bradmcevoy.http.Request;
 import com.bradmcevoy.http.Request.Method;
-import com.bradmcevoy.http.Resource;
+import com.bradmcevoy.http.*;
 import com.bradmcevoy.http.exceptions.BadRequestException;
 import com.bradmcevoy.http.exceptions.NotAuthorizedException;
 import com.bradmcevoy.http.http11.auth.DigestResponse;
+import static com.ettrema.context.RequestContext._;
 import com.ettrema.forms.FormProcessor;
 import com.ettrema.web.component.ComponentDef;
 import com.ettrema.web.component.ComponentUtils;
@@ -24,37 +19,36 @@ import java.util.Collection;
 import java.util.Date;
 import java.util.Map;
 
-import static com.ettrema.context.RequestContext._;
-
 /**
- * Subpages defined in templates are inherited from pages which extend the template
- * 
+ * Subpages defined in templates are inherited from pages which extend the
+ * template
+ *
  * When a subpage is located under a page extending its template, the subpage is
  * wrapped in one of these. The WrappedSubPage records the parent which has
  * inherited the subpage and exposes it as the parent
- * 
- * This wrapping page delegates to the wrapped page for most operations. Its behaviour
- * is that of the subpage, but with the logical parent substituted for the physical
- * parent of the subpage
- * 
+ *
+ * This wrapping page delegates to the wrapped page for most operations. Its
+ * behaviour is that of the subpage, but with the logical parent substituted for
+ * the physical parent of the subpage
+ *
  * @author brad
  */
 public class WrappedSubPage extends CommonTemplated implements PostableResource, ISubPage, Replaceable {
+
     private static final long serialVersionUID = 1L;
-    private static org.apache.log4j.Logger log = org.apache.log4j.Logger.getLogger( WrappedSubPage.class );
-    
-    /**The parent that this isntance was found under. Not the physical parent
-     * of the subpage
-     * 
+    private static org.apache.log4j.Logger log = org.apache.log4j.Logger.getLogger(WrappedSubPage.class);
+    /**
+     * The parent that this isntance was found under. Not the physical parent of
+     * the subpage
+     *
      */
     final Templatable actualParent;
     final ISubPage subPage;
-    
-    
+
     public WrappedSubPage(ISubPage subPage, Templatable actualParent) {
         this.subPage = subPage;
         this.actualParent = actualParent;
-        setContentType( subPage.getContentType(null));
+        setContentType(subPage.getContentType(null));
     }
 
     @Override
@@ -62,61 +56,56 @@ public class WrappedSubPage extends CommonTemplated implements PostableResource,
         return subPage.getDefaultContentType();
     }
 
-
-
     @Override
-    public Object authenticate( DigestResponse digestRequest ) {
-        ClydeAuthenticator authenticator = requestContext().get( ClydeAuthenticator.class );
-        Object o = authenticator.authenticate( actualParent, digestRequest );
-        if( o == null ) {
+    public Object authenticate(DigestResponse digestRequest) {
+        ClydeAuthenticator authenticator = requestContext().get(ClydeAuthenticator.class);
+        Object o = authenticator.authenticate(actualParent, digestRequest);
+        if (o == null) {
             log.warn("authentication failed by: " + authenticator.getClass());
         }
         return o;
     }
 
     @Override
-    public IUser authenticate( String user, String password ) {
-        if( log.isTraceEnabled()) {
-            log.trace( "authenticate(basic): actualParent: " + actualParent.getHref());
+    public IUser authenticate(String user, String password) {
+        if (log.isTraceEnabled()) {
+            log.trace("authenticate(basic): actualParent: " + actualParent.getHref());
         }
-        ClydeAuthenticator authenticator = requestContext().get( ClydeAuthenticator.class );
-        IUser o = authenticator.authenticate( actualParent, user, password );
-        if( o == null ) {
+        ClydeAuthenticator authenticator = requestContext().get(ClydeAuthenticator.class);
+        IUser o = authenticator.authenticate(actualParent, user, password);
+        if (o == null) {
             log.warn("authentication failed by: " + authenticator.getClass());
         }
         return o;
     }
 
     @Override
-    public boolean authorise( Request request, Method method, Auth auth ) {
-        if( this.subPage.isSecure()) {
-            if( auth == null ) {
-                log.debug( "authorisation declined, because subpage is secure, and there is no current user");
+    public boolean authorise(Request request, Method method, Auth auth) {
+        if (this.subPage.isSecure()) {
+            if (auth == null) {
+                log.debug("authorisation declined, because subpage is secure, and there is no current user");
                 return false;
             }
         }
-        if( subPage.isPublicAccess() ) {
+        if (subPage.isPublicAccess()) {
             log.trace("allow access because subpage has public=true");
             return true;
         }
-        return actualParent.authorise( request, method, auth );
+        return actualParent.authorise(request, method, auth);
 
         // Invitations failed authorisation, because authorisation was delegated
         // to the physical resource which do not allow anonymous access
         //return super.authorise( request, method, auth );
     }
 
-
-
     @Override
     public String getUniqueId() {
         return null;
         //return actualParent.getUniqueId() + subPage.getUniqueId();
     }
-    
-    
+
     /**
-     * 
+     *
      * @return - the physical subpage this is wrapping
      */
     public ISubPage getSubPage() {
@@ -124,49 +113,47 @@ public class WrappedSubPage extends CommonTemplated implements PostableResource,
     }
 
     @Override
-    public String getContentType( String accepts ) {
-        return subPage.getContentType( accepts );
+    public String getContentType(String accepts) {
+        return subPage.getContentType(accepts);
     }
 
-
-
     /**
-     * Recursively looks for the physical page under this and subsequent wrappers
-     * 
+     * Recursively looks for the physical page under this and subsequent
+     * wrappers
+     *
      * @return
      */
     public SubPage unwrap() {
-        if( subPage instanceof WrappedSubPage ) {
+        if (subPage instanceof WrappedSubPage) {
             WrappedSubPage next = (WrappedSubPage) subPage;
             return next.unwrap();
         } else {
             return (SubPage) subPage;
         }
     }
-    
+
     @Override
     public Resource getChildResource(String childName) {
         Resource r = null;
         r = subPage.getChildResource(childName);
-        if( r == null ) {
+        if (r == null) {
             r = super.getChildResource(childName);
         }
-        if( r instanceof SubPage ) {
-            return new WrappedSubPage((SubPage) r,this);
+        if (r instanceof SubPage) {
+            return new WrappedSubPage((SubPage) r, this);
         } else {
             return r;
         }
     }
-    
+
     /**
      * For backwards compatibility, same as getParent
-     * 
+     *
      * @return
      */
     public Templatable getFoundParent() {
         return getParent();
     }
-    
 
     @Override
     public Web getWeb() {
@@ -210,35 +197,36 @@ public class WrappedSubPage extends CommonTemplated implements PostableResource,
 
     @Override
     public String processForm(Map<String, String> parameters, Map<String, FileItem> files) throws NotAuthorizedException {
-		return _(FormProcessor.class).processForm(this, parameters, files);
+        return _(FormProcessor.class).processForm(this, parameters, files);
     }
-    
-    /** Components should read their values from request params
+
+    /**
+     * Components should read their values from request params
      */
     @Override
-    public void preProcess( RenderContext rcChild,Map<String, String> parameters, Map<String, FileItem> files ) {
+    public void preProcess(RenderContext rcChild, Map<String, String> parameters, Map<String, FileItem> files) {
         ITemplate lTemplate = getTemplate();
-        RenderContext rc = new RenderContext(lTemplate,this,rcChild,false);
-        if( lTemplate != null ) {
-            lTemplate.preProcess(rc,parameters,files);
-            for( ComponentDef def : lTemplate.getComponentDefs().values() ) {
-                if( !this.getValues().containsKey(def.getName())) {
-                    ComponentValue cv = def.createComponentValue( this );
-                    this.getValues().add( cv );
+        RenderContext rc = new RenderContext(lTemplate, this, rcChild, false);
+        if (lTemplate != null) {
+            lTemplate.preProcess(rc, parameters, files);
+            for (ComponentDef def : lTemplate.getComponentDefs().values()) {
+                if (!this.getValues().containsKey(def.getName())) {
+                    ComponentValue cv = def.createComponentValue(this);
+                    this.getValues().add(cv);
                 }
             }
         }
 
         Collection<Component> all = allComponents();
-        for( Component c : all ) {
+        for (Component c : all) {
             c.init(this);
-            c.onPreProcess(rc,parameters,files);
+            c.onPreProcess(rc, parameters, files);
         }
     }
 
     @Override
-    public void sendContent( OutputStream out, Range range, Map<String, String> params, String contentType ) throws IOException, NotAuthorizedException, BadRequestException {
-        subPage.sendContent( this, out, range, params, contentType );
+    public void sendContent(OutputStream out, Range range, Map<String, String> params, String contentType) throws IOException, NotAuthorizedException, BadRequestException {
+        subPage.sendContent(this, out, range, params, contentType);
     }
 
     /**
@@ -253,39 +241,41 @@ public class WrappedSubPage extends CommonTemplated implements PostableResource,
      * @throws NotAuthorizedException
      * @throws BadRequestException
      */
-	@Override
-    public void sendContent( WrappedSubPage requestedPage, OutputStream out, Range range, Map<String, String> params, String contentType ) throws IOException, NotAuthorizedException, BadRequestException {
-        subPage.sendContent( this, out, range, params, contentType );
+    @Override
+    public void sendContent(WrappedSubPage requestedPage, OutputStream out, Range range, Map<String, String> params, String contentType) throws IOException, NotAuthorizedException, BadRequestException {
+        subPage.sendContent(this, out, range, params, contentType);
     }
 
-
-
-    
-    /** Commands should be invoked, if user clicked
+    /**
+     * Commands should be invoked, if user clicked
      */
     @Override
-    public String process( RenderContext rcChild,Map<String, String> parameters, Map<String, FileItem> files ) throws NotAuthorizedException {
+    public String process(RenderContext rcChild, Map<String, String> parameters, Map<String, FileItem> files) throws NotAuthorizedException {
         ITemplate lTemplate = getTemplate();
-        RenderContext rc = new RenderContext(lTemplate,this,rcChild,false);
+        RenderContext rc = new RenderContext(lTemplate, this, rcChild, false);
         String redirectTo = null;
-        if( lTemplate != null ) {
-            redirectTo = lTemplate.process(rc,parameters,files);
-            if( redirectTo != null ) return redirectTo;
+        if (lTemplate != null) {
+            redirectTo = lTemplate.process(rc, parameters, files);
+            if (redirectTo != null) {
+                return redirectTo;
+            }
         }
-        for( Component c : allComponents() ) {
-            redirectTo = c.onProcess(rc,parameters,files);
-            if( redirectTo != null ) return redirectTo;
+        for (Component c : allComponents()) {
+            redirectTo = c.onProcess(rc, parameters, files);
+            if (redirectTo != null) {
+                return redirectTo;
+            }
         }
         return null;
-    }    
+    }
 
     @Override
     public Collection<Component> allComponents() {
         return ComponentUtils.allComponents(this);
     }
-    
+
     private BaseResource physicalParent(Templatable parentPage) {
-        if( parentPage instanceof BaseResource ) {
+        if (parentPage instanceof BaseResource) {
             return (BaseResource) parentPage;
         }
         return physicalParent(parentPage.getParent());
@@ -293,7 +283,7 @@ public class WrappedSubPage extends CommonTemplated implements PostableResource,
 
     @Override
     public Date getCreateDate() {
-        return physicalParent( subPage ).getCreateDate();
+        return physicalParent(subPage).getCreateDate();
     }
 
     @Override
@@ -301,26 +291,24 @@ public class WrappedSubPage extends CommonTemplated implements PostableResource,
         return subPage.isSecure();
     }
 
-	@Override
+    @Override
     public boolean isPublicAccess() {
         return subPage.isPublicAccess();
     }
 
-
-
-	@Override
+    @Override
     public String getRedirect() {
         return subPage.getRedirect();
     }
 
-	@Override
-    public void replaceContent(InputStream in, Long length) throws BadRequestException{
+    @Override
+    public void replaceContent(InputStream in, Long length) throws BadRequestException, NotAuthorizedException {
         log.trace("replaceContent1");
-        subPage.replaceContent(this,in,length);
+        subPage.replaceContent(this, in, length);
     }
 
-	@Override
-    public void replaceContent(WrappedSubPage requestedPage, InputStream in, Long length) throws BadRequestException {
+    @Override
+    public void replaceContent(WrappedSubPage requestedPage, InputStream in, Long length) throws BadRequestException, NotAuthorizedException {
         log.trace("replaceContent2");
         subPage.replaceContent(this, in, length);
     }
