@@ -23,88 +23,97 @@ import org.apache.commons.collections.CollectionUtils;
  */
 public class LoginResponseHandler extends AbstractWrappingResponseHandler {
 
-    private static org.apache.log4j.Logger log = org.apache.log4j.Logger.getLogger( LoginResponseHandler.class );
+    private static org.apache.log4j.Logger log = org.apache.log4j.Logger.getLogger(LoginResponseHandler.class);
     private String loginPage = "/login.html";
     private ResourceFactory resourceFactory;
     private List<String> excludePaths;
+    private boolean enabled;
 
-    public LoginResponseHandler( WebDavResponseHandler wrapped ) {
-        super( wrapped );
+    public LoginResponseHandler(WebDavResponseHandler wrapped) {
+        super(wrapped);
     }
 
     /**
-     * If responding with a login page, the request attribute "authReason" is set
-     * to either "required", indicating that the user must login; or "notPermitted"
-     * indicating that the user is currently logged in but does not have permission
+     * If responding with a login page, the request attribute "authReason" is
+     * set to either "required", indicating that the user must login; or
+     * "notPermitted" indicating that the user is currently logged in but does
+     * not have permission
      *
      * @param resource
      * @param response
      * @param request
      */
     @Override
-    public void respondUnauthorised( Resource resource, Response response, Request request ) {
+    public void respondUnauthorised(Resource resource, Response response, Request request) {
         log.trace("respondUnauthorised");
         String ctHeader = request.getContentTypeHeader();
-        if( isPage( resource, ctHeader ) && !excluded( request ) && isGetOrPost( request ) ) {
-            Resource rLogin = resourceFactory.getResource( request.getHostHeader(), loginPage );
-            if( rLogin == null || !( rLogin instanceof GetableResource ) ) {
-                log.trace( "Couldnt find login resource: " + request.getHostHeader() + "/" + loginPage );
-                wrapped.respondUnauthorised( resource, response, request );
+        if (isEnabled() && isPage(resource, ctHeader) && !excluded(request) && isGetOrPost(request)) {
+            Resource rLogin;
+            try {
+                rLogin = resourceFactory.getResource(request.getHostHeader(), loginPage);
+            } catch (NotAuthorizedException ex) {
+                throw new RuntimeException(ex);
+            } catch (BadRequestException ex) {
+                throw new RuntimeException(ex);
+            }
+            if (rLogin == null || !(rLogin instanceof GetableResource)) {
+                log.trace("Couldnt find login resource: " + request.getHostHeader() + "/" + loginPage);
+                wrapped.respondUnauthorised(resource, response, request);
             } else {
-                log.trace( "respond with 200 to suppress login prompt" );
+                log.trace("respond with 200 to suppress login prompt, using resource: " + rLogin.getName() + " - " + rLogin.getClass());
                 try {
-                    response.setStatus( Response.Status.SC_OK );
+                    response.setStatus(Response.Status.SC_OK);
                     response.setCacheControlNoCacheHeader();
                     GetableResource gr = (GetableResource) rLogin;
 
                     // set request attribute so rendering knows it authorisation failed, or authentication is required
                     Auth auth = request.getAuthorization();
-                    if( auth != null && auth.getTag() != null ) {
+                    if (auth != null && auth.getTag() != null) {
                         // no authentication was attempted,
-                        request.getAttributes().put( "authReason", "notPermitted" );
+                        request.getAttributes().put("authReason", "notPermitted");
                     } else {
-                        request.getAttributes().put( "authReason", "required" );
+                        request.getAttributes().put("authReason", "required");
                     }
-                    gr.sendContent( response.getOutputStream(), null, null, "text/html" );
+                    gr.sendContent(response.getOutputStream(), null, null, "text/html");
                 } catch (NotFoundException ex) {
-					throw new RuntimeException( ex );
-				} catch( IOException ex ) {
-                    throw new RuntimeException( ex );
-                } catch( NotAuthorizedException ex ) {
-                    throw new RuntimeException( ex );
-                } catch( BadRequestException ex ) {
-                    throw new RuntimeException( ex );
+                    throw new RuntimeException(ex);
+                } catch (IOException ex) {
+                    throw new RuntimeException(ex);
+                } catch (NotAuthorizedException ex) {
+                    throw new RuntimeException(ex);
+                } catch (BadRequestException ex) {
+                    throw new RuntimeException(ex);
                 }
             }
-        } else if ( resource instanceof AjaxPostResource && !excluded( request ) && isGetOrPost( request )) {
+        } else if (resource instanceof AjaxPostResource && !excluded(request) && isGetOrPost(request)) {
             log.trace("ajax post resource, so suppress login prompt");
-            wrapped.respondForbidden( resource, response, request );
+            wrapped.respondForbidden(resource, response, request);
         } else {
-            log.trace( "respond with normal 401" );
-            wrapped.respondUnauthorised( resource, response, request );
+            log.trace("respond with normal 401");
+            wrapped.respondUnauthorised(resource, response, request);
         }
     }
 
-    private boolean isPage( Resource resource, String ctHeader ) {
-        if( resource instanceof GetableResource ) {
+    private boolean isPage(Resource resource, String ctHeader) {
+        if (resource instanceof GetableResource) {
             GetableResource gr = (GetableResource) resource;
-            String ctResource = gr.getContentType( "text/html" );
-            if( ctResource == null ) {
-                if( ctHeader != null ) {
-                    boolean b = ctHeader.contains( "html" );
-                    log.trace( "isPage: resource has no content type, depends on requested content type: " + b );
+            String ctResource = gr.getContentType("text/html");
+            if (ctResource == null) {
+                if (ctHeader != null) {
+                    boolean b = ctHeader.contains("html");
+                    log.trace("isPage: resource has no content type, depends on requested content type: " + b);
                     return b;
                 } else {
-                    log.trace( "isPage: resource has no content type, and no requeted content type, so assume false" );
+                    log.trace("isPage: resource has no content type, and no requeted content type, so assume false");
                     return false;
                 }
             } else {
-                boolean b = ctResource.contains( "html" );
-                log.trace( "isPage: resource has content type. is html? " + b );
+                boolean b = ctResource.contains("html");
+                log.trace("isPage: resource has content type. is html? " + b);
                 return b;
             }
         } else {
-            log.trace( "isPage: resource is not getable" );
+            log.trace("isPage: resource is not getable");
             return false;
         }
     }
@@ -113,7 +122,7 @@ public class LoginResponseHandler extends AbstractWrappingResponseHandler {
         return loginPage;
     }
 
-    public void setLoginPage( String loginPage ) {
+    public void setLoginPage(String loginPage) {
         this.loginPage = loginPage;
     }
 
@@ -121,7 +130,7 @@ public class LoginResponseHandler extends AbstractWrappingResponseHandler {
         return resourceFactory;
     }
 
-    public void setResourceFactory( ResourceFactory resourceFactory ) {
+    public void setResourceFactory(ResourceFactory resourceFactory) {
         this.resourceFactory = resourceFactory;
     }
 
@@ -129,23 +138,32 @@ public class LoginResponseHandler extends AbstractWrappingResponseHandler {
         return excludePaths;
     }
 
-    public void setExcludePaths( List<String> excludePaths ) {
+    public void setExcludePaths(List<String> excludePaths) {
         this.excludePaths = excludePaths;
     }
 
-    private boolean excluded( Request request ) {
-        if( CollectionUtils.isEmpty( excludePaths ) ) {
+    private boolean excluded(Request request) {
+        if (CollectionUtils.isEmpty(excludePaths)) {
             return false;
         }
-        for( String s : excludePaths ) {
-            if( request.getAbsolutePath().startsWith( s ) ) {
+        for (String s : excludePaths) {
+            if (request.getAbsolutePath().startsWith(s)) {
                 return true;
             }
         }
         return false;
     }
 
-    private boolean isGetOrPost( Request request ) {
-        return request.getMethod().equals( Method.GET ) || request.getMethod().equals( Method.POST );
+    private boolean isGetOrPost(Request request) {
+        return request.getMethod().equals(Method.GET) || request.getMethod().equals(Method.POST);
     }
+
+    public boolean isEnabled() {
+        return enabled;
+    }
+
+    public void setEnabled(boolean enabled) {
+        this.enabled = enabled;
+    }
+        
 }
