@@ -1,26 +1,14 @@
 package com.ettrema.facebook;
 
 import com.bradmcevoy.common.Path;
-import com.bradmcevoy.http.Auth;
-import com.bradmcevoy.http.Cookie;
-import com.bradmcevoy.http.FileItem;
-import com.bradmcevoy.http.GetableResource;
-import com.bradmcevoy.http.PostableResource;
-import com.bradmcevoy.http.Range;
-import com.bradmcevoy.http.Request;
 import com.bradmcevoy.http.Request.Method;
-import com.bradmcevoy.http.Resource;
-import com.bradmcevoy.http.ResourceFactory;
+import com.bradmcevoy.http.*;
 import com.bradmcevoy.http.exceptions.BadRequestException;
 import com.bradmcevoy.http.exceptions.ConflictException;
 import com.bradmcevoy.http.exceptions.NotAuthorizedException;
 import com.bradmcevoy.http.exceptions.NotFoundException;
 import com.bradmcevoy.io.BufferingOutputStream;
-import com.restfb.Connection;
-import com.restfb.DefaultFacebookClient;
-import com.restfb.FacebookClient;
-import com.restfb.FacebookException;
-import com.restfb.Parameter;
+import com.restfb.*;
 import com.restfb.types.Album;
 import com.restfb.types.FacebookType;
 import java.io.IOException;
@@ -40,30 +28,30 @@ import org.apache.commons.io.IOUtils;
  */
 public class FacebookResourceFactory implements ResourceFactory {
 
-    private static org.apache.log4j.Logger log = org.apache.log4j.Logger.getLogger( FacebookResourceFactory.class );
+    private static org.apache.log4j.Logger log = org.apache.log4j.Logger.getLogger(FacebookResourceFactory.class);
     private final ResourceFactory wrapped;
     private String name = "_sys_facebook";
     private String appId;
     private String appSecret;
 
-    public FacebookResourceFactory( ResourceFactory wrapped ) {
+    public FacebookResourceFactory(ResourceFactory wrapped) {
         this.wrapped = wrapped;
     }
 
-	@Override
-    public Resource getResource( String host, String sPath ) {
-        log.trace( "getResource" );
-        Path path = Path.path( sPath );
-        if( path.getName().equals( name ) ) {
-            Resource toSend = wrapped.getResource( host, path.getParent().toString() );
-            if( toSend == null ) {
+    @Override
+    public Resource getResource(String host, String sPath) throws NotAuthorizedException, BadRequestException {
+        log.trace("getResource");
+        Path path = Path.path(sPath);
+        if (path.getName().equals(name)) {
+            Resource toSend = wrapped.getResource(host, path.getParent().toString());
+            if (toSend == null) {
                 return null;
             } else {
-                if( toSend instanceof GetableResource ) {
-                    log.trace( "got resource" );
-                    return new UploadResource( (GetableResource) toSend, path.getParent() );
+                if (toSend instanceof GetableResource) {
+                    log.trace("got resource");
+                    return new UploadResource((GetableResource) toSend, path.getParent());
                 } else {
-                    log.warn( "not compatible resource: " + toSend.getClass().getCanonicalName() );
+                    log.warn("not compatible resource: " + toSend.getClass().getCanonicalName());
                     return null;
                 }
             }
@@ -74,49 +62,51 @@ public class FacebookResourceFactory implements ResourceFactory {
 
     /**
      * return the access token
-     * 
+     *
      * @param request
      * @return
      */
-    private Map<String, String> parseRequest( Request request ) throws NoSuchAlgorithmException {
+    private Map<String, String> parseRequest(Request request) throws NoSuchAlgorithmException {
         String cookieName = "fbs_" + appId;
-        log.debug( "parseRequest: " + cookieName );
-        Cookie cookie = request.getCookie( cookieName );
-        if( cookie == null ) {
-            log.debug( "no fb cookie" );
+        log.debug("parseRequest: " + cookieName);
+        Cookie cookie = request.getCookie(cookieName);
+        if (cookie == null) {
+            log.debug("no fb cookie");
             return null;
         }
         String fbCookieValue = cookie.getValue();
-        log.debug( "fbCookieValue: " + fbCookieValue );
+        log.debug("fbCookieValue: " + fbCookieValue);
 
         //remove first and last double quotes
-        String[] pairs = fbCookieValue.split( "\"|&" );
-        Arrays.sort( pairs );
+        String[] pairs = fbCookieValue.split("\"|&");
+        Arrays.sort(pairs);
         StringBuilder payload = new StringBuilder();
         Map<String, String> map = new HashMap<String, String>();
-        for( int i = 0; i < pairs.length; i++ ) {
+        for (int i = 0; i < pairs.length; i++) {
             String pair = pairs[i];
-            if( !( pair.contains( "sig=" ) ) ) {
-                payload.append( pair );
+            if (!(pair.contains("sig="))) {
+                payload.append(pair);
             }
-            String[] nv = pair.split( "=" );
-            map.put( nv[0], nv[1] );
+            String[] nv = pair.split("=");
+            map.put(nv[0], nv[1]);
         }
-        payload.append( appSecret );
-        log.debug( "payload: " + payload );
+        payload.append(appSecret);
+        log.debug("payload: " + payload);
 
-        MessageDigest digest = java.security.MessageDigest.getInstance( "MD5" );
+        MessageDigest digest = java.security.MessageDigest.getInstance("MD5");
         digest.reset();
-        digest.update( payload.toString().getBytes() );
+        digest.update(payload.toString().getBytes());
         byte[] hash = digest.digest();
         StringBuilder buf = new StringBuilder();
 
-        for( int i = 0; i < hash.length; i++ ) {
-            String hex = Integer.toHexString( 0xff & hash[i] );
-            if( hex.length() == 1 ) buf.append( '0' );
-            buf.append( hex );
+        for (int i = 0; i < hash.length; i++) {
+            String hex = Integer.toHexString(0xff & hash[i]);
+            if (hex.length() == 1) {
+                buf.append('0');
+            }
+            buf.append(hex);
         }
-        log.debug( "hash: " + buf );
+        log.debug("hash: " + buf);
         return map;
     }
 
@@ -124,7 +114,7 @@ public class FacebookResourceFactory implements ResourceFactory {
         return name;
     }
 
-    public void setName( String name ) {
+    public void setName(String name) {
         this.name = name;
     }
 
@@ -132,7 +122,7 @@ public class FacebookResourceFactory implements ResourceFactory {
         return appId;
     }
 
-    public void setAppId( String appId ) {
+    public void setAppId(String appId) {
         this.appId = appId;
     }
 
@@ -140,7 +130,7 @@ public class FacebookResourceFactory implements ResourceFactory {
         return appSecret;
     }
 
-    public void setAppSecret( String appSecret ) {
+    public void setAppSecret(String appSecret) {
         this.appSecret = appSecret;
     }
 
@@ -151,54 +141,54 @@ public class FacebookResourceFactory implements ResourceFactory {
         private String accessToken;
         private String result;
 
-        public UploadResource( GetableResource toSend, Path resourcePath ) {
+        public UploadResource(GetableResource toSend, Path resourcePath) {
             this.toSend = toSend;
             this.resourcePath = resourcePath;
         }
 
-		@Override
-        public String processForm( Map<String, String> parameters, Map<String, FileItem> files ) throws BadRequestException, NotAuthorizedException, ConflictException {
-            log.debug( "processForm" );
-            if( accessToken == null ) {
-                throw new BadRequestException( this );
+        @Override
+        public String processForm(Map<String, String> parameters, Map<String, FileItem> files) throws BadRequestException, NotAuthorizedException, ConflictException {
+            log.debug("processForm");
+            if (accessToken == null) {
+                throw new BadRequestException(this);
             }
             try {
-                FacebookClient facebookClient = new DefaultFacebookClient( accessToken );
+                FacebookClient facebookClient = new DefaultFacebookClient(accessToken);
                 String destAlbum = getAlbumName();
-                log.debug( "destAlbum: " + destAlbum );
-                String destAlbumId = getOrCreateAlbum( facebookClient, destAlbum );
-                log.debug( "destAlbumId: " + destAlbumId );
-                String caption = parameters.get( "caption");
-                if(caption == null || caption.length() == 0 ) {
+                log.debug("destAlbum: " + destAlbum);
+                String destAlbumId = getOrCreateAlbum(facebookClient, destAlbum);
+                log.debug("destAlbumId: " + destAlbumId);
+                String caption = parameters.get("caption");
+                if (caption == null || caption.length() == 0) {
                     caption = "Published by shmego.com";
                 }
                 InputStream content = null;
                 try {
                     content = getContent();
                     destAlbumId = destAlbumId + "/photos";
-                    FacebookType publishPhotoResponse = facebookClient.publish( destAlbumId, FacebookType.class, content, Parameter.with( "message", caption ) );
-                    log.debug( "publishjed ok: " + publishPhotoResponse.getId() );
+                    FacebookType publishPhotoResponse = facebookClient.publish(destAlbumId, FacebookType.class, content, Parameter.with("message", caption));
+                    log.debug("publishjed ok: " + publishPhotoResponse.getId());
                 } catch (NotFoundException ex) {
-					throw new RuntimeException( ex );
-				} catch( IOException ex ) {
-                    throw new RuntimeException( ex );
+                    throw new RuntimeException(ex);
+                } catch (IOException ex) {
+                    throw new RuntimeException(ex);
                 } finally {
-                    IOUtils.closeQuietly( content );
+                    IOUtils.closeQuietly(content);
                 }
                 result = "ok";
-            } catch( FacebookException ex ) {
-                log.error( "exception loading: " + resourcePath, ex );
+            } catch (FacebookException ex) {
+                log.error("exception loading: " + resourcePath, ex);
                 result = "err";
             }
             return null;
         }
 
-		@Override
-        public void sendContent( OutputStream out, Range range, Map<String, String> params, String contentType ) throws IOException, NotAuthorizedException, BadRequestException {
+        @Override
+        public void sendContent(OutputStream out, Range range, Map<String, String> params, String contentType) throws IOException, NotAuthorizedException, BadRequestException {
             String s;
-            if( result == null ) {
+            if (result == null) {
                 // not a POST
-                if( accessToken != null ) {
+                if (accessToken != null) {
                     s = "loggedIn";
                 } else {
                     s = "notLoggedIn";
@@ -206,14 +196,14 @@ public class FacebookResourceFactory implements ResourceFactory {
             } else {
                 s = result;
             }
-            out.write( s.getBytes() );
+            out.write(s.getBytes());
         }
 
-        public Long getMaxAgeSeconds( Auth auth ) {
+        public Long getMaxAgeSeconds(Auth auth) {
             return null;
         }
 
-        public String getContentType( String accepts ) {
+        public String getContentType(String accepts) {
             return "text/plain";
         }
 
@@ -229,63 +219,67 @@ public class FacebookResourceFactory implements ResourceFactory {
             return name;
         }
 
-        public Object authenticate( String user, String password ) {
-            return toSend.authenticate( user, password );
+        public Object authenticate(String user, String password) {
+            return toSend.authenticate(user, password);
         }
 
-        public boolean authorise( Request request, Method method, Auth auth ) {
-            return toSend.authorise( request, method, auth );
+        public boolean authorise(Request request, Method method, Auth auth) {
+            return toSend.authorise(request, method, auth);
         }
 
         public String getRealm() {
             return toSend.getRealm();
         }
 
+        @Override
         public Date getModifiedDate() {
             return null;
         }
 
-        public String checkRedirect( Request request ) {
+        @Override
+        public String checkRedirect(Request request) {
             try {
-                Map<String, String> vals = parseRequest( request );
-                if( vals != null ) {
-                    accessToken = vals.get( "access_token" );
+                Map<String, String> vals = parseRequest(request);
+                if (vals != null) {
+                    accessToken = vals.get("access_token");
                 }
                 return null;
-            } catch( NoSuchAlgorithmException ex ) {
-                throw new RuntimeException( ex );
+            } catch (NoSuchAlgorithmException ex) {
+                throw new RuntimeException(ex);
             }
         }
 
         private String getAlbumName() {
             Path p = resourcePath.getParent();
-            if( p.getName().startsWith( "_sys")) p = p.getParent();
+            if (p.getName().startsWith("_sys")) {
+                p = p.getParent();
+            }
             return p.getName();
         }
 
-        private String getOrCreateAlbum( FacebookClient facebookClient, String destAlbum ) throws FacebookException {
-            Connection<Album> myAlbums = facebookClient.fetchConnection( "me/albums", Album.class );
-            for( Album a : myAlbums.getData() ) {
-                if( a.getName().equals( destAlbum ) ) {
-                    log.debug( "got album: " + a.getId() );
+        private String getOrCreateAlbum(FacebookClient facebookClient, String destAlbum) throws FacebookException {
+            Connection<Album> myAlbums = facebookClient.fetchConnection("me/albums", Album.class);
+            for (Album a : myAlbums.getData()) {
+                if (a.getName().equals(destAlbum)) {
+                    log.debug("got album: " + a.getId());
                     return a.getId();
                 }
             }
-            log.debug( "does not exist, create it" );
-            FacebookType resp = facebookClient.publish( "/me/albums", FacebookType.class, Parameter.with( "name", destAlbum ) );
-            log.debug( "created: " + resp.getId() );
+            log.debug("does not exist, create it");
+            FacebookType resp = facebookClient.publish("/me/albums", FacebookType.class, Parameter.with("name", destAlbum));
+            log.debug("created: " + resp.getId());
             return resp.getId();
         }
 
         private InputStream getContent() throws IOException, NotAuthorizedException, BadRequestException, NotFoundException {
             BufferingOutputStream out = null;
             try {
-                out = new BufferingOutputStream( 50000 );
-                toSend.sendContent( out, null, null, null );
+                out = new BufferingOutputStream(50000);
+                toSend.sendContent(out, null, null, null);
             } finally {
-                IOUtils.closeQuietly( out );
+                IOUtils.closeQuietly(out);
             }
-            log.debug( "data: " + out.getSize() );
+            log.debug("data: " + out.getSize());
             return out.getInputStream();
         }
     }
