@@ -1,13 +1,7 @@
 package com.ettrema.web.manage.synch;
 
-import java.util.Date;
 import com.bradmcevoy.common.ContentTypeUtils;
-import com.bradmcevoy.http.ReplaceableResource;
-import com.bradmcevoy.http.CollectionResource;
-import com.bradmcevoy.http.DeletableResource;
-import com.bradmcevoy.http.MakeCollectionableResource;
-import com.bradmcevoy.http.PutableResource;
-import com.bradmcevoy.http.Resource;
+import com.bradmcevoy.http.*;
 import com.bradmcevoy.http.exceptions.BadRequestException;
 import com.bradmcevoy.http.exceptions.ConflictException;
 import com.bradmcevoy.http.exceptions.NotAuthorizedException;
@@ -16,42 +10,43 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.util.Date;
 import org.apache.commons.io.IOUtils;
-
 
 /**
  *
  * @author brad
  */
-public class DirectFileTransport  implements FileTransport{
+public class DirectFileTransport implements FileTransport {
 
-	private static org.apache.log4j.Logger log = org.apache.log4j.Logger.getLogger(FileLoader.class);
-	
-	private final CodeResourceFactory resourceFactory;
-	
-	private String hostName = "test.ppod.com";  // todo: fix config to use localhost
+    private static org.apache.log4j.Logger log = org.apache.log4j.Logger.getLogger(DirectFileTransport.class);
+    private final CodeResourceFactory resourceFactory;
+    private final FileLoadCallback callback;
+    private String hostName = "localhost";
 
-	public DirectFileTransport(String hostName, CodeResourceFactory resourceFactory) {
-		this.resourceFactory = resourceFactory;
-		this.hostName = hostName;
-	}
-	
-	public DirectFileTransport(CodeResourceFactory resourceFactory) {
-		this.resourceFactory = resourceFactory;
-		this.hostName = "test.com";
-	}
-	
-	
+    public DirectFileTransport(String hostName, CodeResourceFactory resourceFactory) {
+        this.resourceFactory = resourceFactory;
+        this.hostName = hostName;
+        this.callback = null;
+    }
+
+    public DirectFileTransport(String hostName, CodeResourceFactory resourceFactory, FileLoadCallback callback) {
+        this.resourceFactory = resourceFactory;
+        this.hostName = hostName;
+        this.callback = callback;
+    }
+
     /**
-     * Just upload the given file to its parent directory.
-     * No name transformations
+     * Just upload the given file to its parent directory. No name
+     * transformations
+     *
      * @param f
      */
-	@Override
+    @Override
     public void put(File f, File root) throws NotAuthorizedException, ConflictException, BadRequestException, IOException {
-        log.trace("put: " + f.getAbsolutePath());
+        log.info("put: " + f.getAbsolutePath() + " root: " + root.getAbsolutePath());
         CollectionResource colParent = findCollection(f.getParentFile(), root);
-        if( colParent == null ) {
+        if (colParent == null) {
             throw new RuntimeException("Couldnt locate parent: " + f.getParentFile().getAbsolutePath() + " for root: " + root.getAbsolutePath());
         }
         Resource rExisting = colParent.child(f.getName());
@@ -62,6 +57,9 @@ public class DirectFileTransport  implements FileTransport{
             try {
                 fin = new FileInputStream(f);
                 replaceable.replaceContent(fin, f.length());
+                if (callback != null) {
+                    callback.onModified(replaceable);
+                }
             } catch (FileNotFoundException ex) {
                 throw new RuntimeException(ex);
             } finally {
@@ -78,7 +76,14 @@ public class DirectFileTransport  implements FileTransport{
                 try {
                     fin = new FileInputStream(f);
                     String ct = ContentTypeUtils.findContentTypes(f);
-                    putable.createNew(f.getName(), fin, f.length(), ct);
+                    Resource newRes = putable.createNew(f.getName(), fin, f.length(), ct);
+                    if (callback != null) {
+                        if (rExisting != null) {
+                            callback.onModified(newRes);
+                        } else {
+                            callback.onLoaded(newRes);
+                        }
+                    }
                 } catch (FileNotFoundException ex) {
                     throw new RuntimeException(ex);
                 } finally {
@@ -121,8 +126,8 @@ public class DirectFileTransport  implements FileTransport{
 
     }
 
-	@Override
-    public  void delete(File f, File root) throws NotAuthorizedException, ConflictException, BadRequestException {
+    @Override
+    public void delete(File f, File root) throws NotAuthorizedException, ConflictException, BadRequestException {
         log.trace("delete: " + f.getAbsolutePath());
         File fMeta = CodeSynchUtils.toMetaFile(f);
         Resource r = resourceFactory.getResource(hostName, CodeSynchUtils.toCodePath(fMeta, root));
@@ -131,16 +136,24 @@ public class DirectFileTransport  implements FileTransport{
         } else if (r instanceof DeletableResource) {
             DeletableResource dr = (DeletableResource) r;
             dr.delete();
+            if (callback != null) {
+                callback.onDeleted(dr);
+            }
         } else {
             throw new RuntimeException("Cannot delete " + f.getAbsolutePath() + " is a : " + r.getClass());
         }
 
-    }	
-	
-	@Override
+    }
+
+    @Override
     public boolean isNewOrUpdated(File f, File root) {
         f = CodeSynchUtils.toMetaFile(f);
-        Resource r = resourceFactory.getResource(hostName, CodeSynchUtils.toCodePath(f, root));
+        Resource r;
+        try {
+            r = resourceFactory.getResource(hostName, CodeSynchUtils.toCodePath(f, root));
+        } catch (NotAuthorizedException | BadRequestException ex) { 
+            throw new RuntimeException(ex);
+        }
         if (r == null || r.getModifiedDate() == null) {
             return true;
         } else {
@@ -148,13 +161,21 @@ public class DirectFileTransport  implements FileTransport{
             return fileModDate.after(r.getModifiedDate());
         }
     }
-	
+
     public String getHostName() {
         return hostName;
     }
 
     public void setHostName(String hostName) {
         this.hostName = hostName;
-    }	
-	
+    }
+
+    public interface FileLoadCallback {
+
+        void onLoaded(Resource replaceable);
+
+        void onDeleted(Resource replaceable);
+
+        public void onModified(Resource replaceable);
+    }
 }
