@@ -3,6 +3,8 @@ package com.ettrema.media;
 import com.bradmcevoy.common.Path;
 import com.bradmcevoy.http.Resource;
 import com.bradmcevoy.http.ResourceFactory;
+import com.bradmcevoy.http.exceptions.BadRequestException;
+import com.bradmcevoy.http.exceptions.NotAuthorizedException;
 import com.ettrema.web.Folder;
 
 /**
@@ -11,7 +13,7 @@ import com.ettrema.web.Folder;
  */
 public class MediaFeedResourceFactory implements ResourceFactory {
 
-    private static org.apache.log4j.Logger log = org.apache.log4j.Logger.getLogger( MediaFeedResourceFactory.class );
+    private static org.apache.log4j.Logger log = org.apache.log4j.Logger.getLogger(MediaFeedResourceFactory.class);
     private final String feedName;
     private final ResourceFactory wrapped;
     private MediaFeedLinkGenerator linkGenerator;
@@ -19,26 +21,26 @@ public class MediaFeedResourceFactory implements ResourceFactory {
     private Long cacheSeconds;
     private boolean secure;
 
-    public MediaFeedResourceFactory( String rssName, ResourceFactory wrapped, MediaLogServiceImpl logService ) {
+    public MediaFeedResourceFactory(String rssName, ResourceFactory wrapped, MediaLogServiceImpl logService) {
         this.feedName = rssName;
         this.wrapped = wrapped;
         this.logService = logService;
     }
 
-	@Override
-    public Resource getResource( String host, String sPath ) {
-        log.trace( "getResource: " + sPath );
-        Path path = Path.path( sPath );
-        if( path.getName().equals( feedName ) ) {
-            log.trace( "got media feed name" );
-            Resource parent = wrapped.getResource( host, path.getParent().toString() );
-            if( parent instanceof Folder ) {
+    @Override
+    public Resource getResource(String host, String sPath) throws NotAuthorizedException, BadRequestException {
+        log.trace("getResource: " + sPath);
+        Path path = Path.path(sPath);
+        if (path.getName().equals(feedName)) {
+            log.trace("got media feed name");
+            Resource parent = wrapped.getResource(host, path.getParent().toString());
+            if (parent instanceof Folder) {
                 Folder folder = (Folder) parent;
                 String basePath = buildBasePath(host, path.getParent());
-                log.trace( "got media feed resource" );
-                return new MediaFeedResource( logService, linkGenerator, feedName, folder, cacheSeconds, basePath );
+                log.trace("got media feed resource");
+                return new MediaFeedResource(logService, linkGenerator, feedName, folder, cacheSeconds, basePath);
             } else {
-                log.trace( "did not find: " + path.getParent() );
+                log.trace("did not find: " + path.getParent());
                 return null;
             }
         } else {
@@ -51,7 +53,7 @@ public class MediaFeedResourceFactory implements ResourceFactory {
         return cacheSeconds;
     }
 
-    public void setCacheSeconds( Long cacheSeconds ) {
+    public void setCacheSeconds(Long cacheSeconds) {
         this.cacheSeconds = cacheSeconds;
     }
 
@@ -59,14 +61,14 @@ public class MediaFeedResourceFactory implements ResourceFactory {
         return secure;
     }
 
-    public void setSecure( boolean secure ) {
+    public void setSecure(boolean secure) {
         this.secure = secure;
     }
 
-    private String buildBasePath( String host, Path parent ) {
+    private String buildBasePath(String host, Path parent) {
         String prot = secure ? "https" : "http";
         String s = prot + "://" + host + parent.toString();
-        log.debug( "base path: " + s);
+        log.debug("base path: " + s);
         return s;
     }
 
@@ -74,10 +76,7 @@ public class MediaFeedResourceFactory implements ResourceFactory {
         return linkGenerator;
     }
 
-    public void setLinkGenerator( MediaFeedLinkGenerator linkGenerator ) {
+    public void setLinkGenerator(MediaFeedLinkGenerator linkGenerator) {
         this.linkGenerator = linkGenerator;
     }
-
-    
-
 }
